WFLY-14449 ChannelCommandDispatcherFactory should not assume JBoss Modules
